misc: altera os links do Comprador para usarem classes utilitarias
